Refactoring to get rid of the two URL's for the environment
