+++ apps/website/src/environments/environment.prod.ts
@@ -1,6 +1,5 @@
 export const environment = {
   production: true,
   saveSettings: true,
-  url: 'http://192.168.2.249',
-  websockUrl: 'https://api.angulon.steinjonker.nl'
+  url: 'https://api.angulon.steinjonker.nl'
 }
